Move Dirlight to lights file with the lights uniform manager type
