+++ source/gfx/render_pipeline.cpp
@@ -4,12 +4,12 @@
 #include <glm/gtc/matrix_transform.hpp>
 
 #include "render_pipeline.h"
-#include "game/transform.h"
 #include "texture.h"
 #include "shader_manager.h"
 #include "texture_manager.h"
 #include "material_manager.h"
 
+#include "light.h"
 #include "camera.h"
 
 namespace gfx {
@@ -26,51 +26,6 @@ namespace gfx {
         static inline gpu::VertexLayout layout;
     };
 
-    class Lights {
-    public:
-        Lights() = default;
-        Lights(const std::vector<gpu::DirLight> &dirLights)
-            : mDirLightsCount(static_cast<int>(dirLights.size()))
-            , mDirLights(dirLights)
-        {
-            mLayout = std::make_unique<gpu::BufferLayout>();
-            mLayout->addAttribute("dirLightsCount", sizeof(int));
-
-            for (int i = 0; i < mDirLightsCount; ++i) {
-                mLayout->addAttribute("dirLights[" + std::to_string(i) + "].direction", sizeof(glm::vec3));
-                mLayout->addAttribute("dirLights[" + std::to_string(i) + "].ambient", sizeof(glm::vec3));
-                mLayout->addAttribute("dirLights[" + std::to_string(i) + "].diffuse", sizeof(glm::vec3));
-                mLayout->addAttribute("dirLights[" + std::to_string(i) + "].specular", sizeof(glm::vec3));
-            }
-
-            mBuffer = gpu::SharedUniformBuffer::create(LightsBlockBinding, mLayout->totalSize());
-        }
-
-        void setBufferData() {
-            auto dirLightsCountOffset = mLayout->attribute("dirLightsCount").offset;
-
-            mBuffer->setData(dirLightsCountOffset, &mDirLightsCount, sizeof(int));
-
-            for (int i = 0; i < mDirLightsCount; i++) {
-                auto directionOffset = mLayout->attribute("dirLights[" + std::to_string(i) + "].direction").offset;
-                auto ambientOffset = mLayout->attribute("dirLights[" + std::to_string(i) + "].ambient").offset;
-                auto diffuseOffset = mLayout->attribute("dirLights[" + std::to_string(i) + "].diffuse").offset;
-                auto specularOffset = mLayout->attribute("dirLights[" + std::to_string(i) + "].specular").offset;
-
-                mBuffer->setData(directionOffset, &mDirLights[i].direction, sizeof(glm::vec3));
-                mBuffer->setData(ambientOffset, &mDirLights[i].ambient, sizeof(glm::vec3));
-                mBuffer->setData(diffuseOffset, &mDirLights[i].diffuse, sizeof(glm::vec3));
-                mBuffer->setData(specularOffset, &mDirLights[i].specular, sizeof(glm::vec3));
-            }
-        }
-    private:
-        std::unique_ptr<gpu::SharedUniformBuffer> mBuffer;
-        std::unique_ptr<gpu::BufferLayout> mLayout;
-
-        int mDirLightsCount;
-        std::vector<gpu::DirLight> mDirLights;
-    };
-
     class RenderPipelineImpl : public RenderPipeline {
     public:
         explicit RenderPipelineImpl(Allocator &allocator, math::Size2D frameDimensions)
@@ -83,8 +38,8 @@ namespace gfx {
         void initialize() override {
             PosTextVertex::init();
 
-            gpu::DirLight dirLight {};
-            std::vector<gpu::DirLight> dirLights;
+            DirLight dirLight {};
+            std::vector<DirLight> dirLights;
 
             dirLight.direction = glm::vec3(0.0f, -1.0f, -1.0f);
             dirLight.ambient = glm::vec3(0.2f, 0.2f, 0.2f);
